cookie로 로그인 인증

+++ app.js
@@ -1,6 +1,8 @@
 // 기본적인 node.js 세팅
 const express = require('express');
 const app = express();
+const http = require('http');
+const cookieParser = require('cookie-parser');
 
 // mysql 세팅
 const db = require('./config/mysql.js');
@@ -8,7 +10,31 @@ const db = require('./config/mysql.js');
 // app 세팅
 app.use(express.json()); // json을 사용하겠다~
 app.use(express.urlencoded({ extended: true }));
+app.use(express.static('public'));
+app.use(cookieParser()); // cookie-parser 미들웨어 사용
 app.set('view engine', 'html');
+
+const middleware = async (req, res, next) => {
+    // 로그인이 되지 않았을 때는 게시판으로 넘어가지 않게 하기
+    let allowedUrls = ['/', '/board/signup']; // 예외 url설정, 메인화면과 회원가입 페이지 => 쿠키가 필요없음
+    let isLoginRequired = true; // 로그인이 필요한지
+
+    for await (let allowedUrl of allowedUrls) {
+        if (req.path == allowedUrl) {
+            // 만약 path == 허락한 url이라면 cookie를 생성해야 함
+            isLoginRequired = false; // false 3처리해주고
+        }
+    }
+
+    if (isLoginRequired && !req.cookies.id) {
+        // 로그인이 필요하면서 쿠키를 생성하려면 메인화면으로
+        return res.redirect('/');
+    }
+    // 그게 아니라면 next때려버리기
+    return next();
+};
+
+app.use(middleware);
 
 // nunjucks 템플릿 설정
 const nunjucks = require('nunjucks'); // 템플릿? 검색하다가 있길래
@@ -16,6 +42,13 @@ nunjucks.configure('views', {
     express: app,
     watch: true,
 });
+
+// http.createServer(function (req, res) {
+//     res.writeHead(200, {
+//         'Set-Cookie': 'c1=cookie1!',
+//         'Content-Type': 'text/html; charset=utf-8',
+//     });
+// });
 
 //게시판 메인 페이지
 app.get('/', function (req, res) {
@@ -34,18 +67,28 @@ app.get('/board/list', async function (req, res) {
     await conn.destroy();
     return res.render('board_list.html', {
         content: result,
+        id: req.cookies.id,
     });
 });
 
 // 글을 쓰는 창
 app.get('/board/write', async function (req, res) {
-    return res.render('board_write.html');
+    // 글을 쓰는 창을 넘어갔을 때 작성자에 로그인한 id가 떠야 됨
+    // 그럴려면 현재 가지고 있는 쿠키의 id값을 가져온다
+    // 그리고 나서 render할때 id값을 같이 넘겨줘서 출력을 하게 한다!
+
+    let id = req.cookies.id;
+    console.log(id);
+
+    return res.render('board_write.html', {
+        content: id,
+    });
 });
 
 // 작성하기를 누르면 알아서 board/list로 넘어가짐
 app.post('/board/write', async function (req, res) {
     // 내가 필요한 정보들 알아오기
-    const username = req.body.username;
+    const username = req.cookies.id;
     const subject = req.body.subject;
     const date = req.body.data;
 
@@ -57,7 +100,7 @@ app.post('/board/write', async function (req, res) {
 
     // query문 안에서 쓰일 파라미터 값
     let values = [subject, username];
-    await conn.query(sql, values);
+    let [result] = await conn.query(sql, values);
     await conn.destroy();
     return res.redirect('/board/list');
 });
@@ -72,14 +115,20 @@ app.get('/board/view', function (req, res) {
 
 //글을 삭제하기
 app.post('/board/delete', async function (req, res) {
+    const index = req.body.idx;
+    const userID = req.cookies.id;
+    let sql = `select idx from Board where username=? and idx=?`;
     const conn = await db.getConnection();
+    let [result] = await conn.query(sql, [userID, index]);
 
-    // 글을 삭제하기 위해서는 고유값인 idx값을 알아와야됨
-    let index = req.body.idx;
-    let sql = `delete from Board where idx=?`;
+    if (result.length === 0) {
+        console.log('삭제 불가능');
+        await conn.destroy();
+        return res.redirect('/board/list');
+    }
 
-    // 마찬가지로 index값을 배열에 넣어서
-    await conn.query(sql, [index]);
+    let deleteSQL = `delete from Board where idx=?`;
+    await conn.query(deleteSQL, [index]);
     await conn.destroy();
     return res.redirect('/board/list');
 });
@@ -103,7 +152,7 @@ app.get('/board/update', async function (req, res) {
     // 응답하는 거는 앞에 return을 붙여주는 것이 좋다
     await conn.destroy();
     return res.render('board_update.html', {
-        // reultr값은 하나만 반환하는 것이 아니므로 배열에 들어가진다, 따라서 인덱싱을 통해 값을 가져올 것
+        // result값은 하나만 반환하는 것이 아니므로 배열에 들어가진다, 따라서 인덱싱을 통해 값을 가져올 것
         content: result[0],
     });
 });
@@ -111,10 +160,17 @@ app.post('/board/update', async function (req, res) {
     // 입력한 값을 받아오기
     const index = req.body.idx;
     const subject = req.body.subject;
-    const username = req.body.username;
     const date = req.body.date;
+    const userID = req.cookies.id;
     const conn = await db.getConnection();
-    console.log(index);
+
+    // 보드 테이블에서 쿠키 id를 통해 idx값을 디져서 같을 때에만 수정할 수 있도록 하고, 아니라면 홈페이지르 돌려버리자
+    let checkSQL = `select idx from Board where username=? and idx=?`;
+    let [checkResult] = await conn.query(checkSQL, [userID, index]);
+    if (checkResult.length === 0) {
+        console.log('수정할 수 없는 접근입니다');
+        return res.redirect('/board/list');
+    }
 
     // db가져올 sql문 작성하기
     let sql = `update Board set username=?, subject=?, date=? where idx=?`;
@@ -132,6 +188,72 @@ app.get('/board/view', function (req, res) {
         data: view,
         index: index,
     });
+});
+app.get('/', async function (req, res) {
+    return res.render('border_list.html');
+});
+// 로그인 창
+app.post('/', async function (req, res) {
+    const id = req.body.id;
+    const password = req.body.password;
+    console.log('로그인 요청');
+    const conn = await db.getConnection(); // db정보 가져오기
+
+    let sql = `select id, email from User where id=? and password=? `;
+    let values = [id, password];
+    let [result] = await conn.query(sql, values);
+    console.log(result);
+    if (result.length > 0) {
+        console.log('로그인 성공');
+        res.cookie('id', id);
+        return res.redirect('/board/list');
+    } else {
+        console.log('로그인 실패');
+        return res.redirect('/');
+    }
+});
+// 로그아웃 창
+app.post('/board/logout', async function (req, res) {
+    let id = req.cookies.id;
+    res.clearCookie('id');
+    return res.redirect('/');
+});
+
+// 회원가입 창 + 회원가입시 쿠키를 만들어주자
+app.post('/board/signup', async function (req, res) {
+    const id = req.body.id;
+    const password = req.body.password;
+    const email = req.body.email;
+    const phoneNumber = req.body.phoneNumber;
+
+    console.log('호출됨' + req);
+    const conn = await db.getConnection(); // db정보 가져오기
+
+    let checkSQL = `select * from User where id=?`;
+    let [result2] = await conn.query(checkSQL, [id]);
+    if (result2.length > 0) {
+        await conn.destroy();
+        console.log('이미 있는 아이디');
+        return res.redirect('/board/signup');
+    } else {
+        let sql = 'insert into User(id,password,email,phoneNumber) values(?,?,?,?) ';
+        let values = [id, password, email, phoneNumber];
+        let [result] = await conn.query(sql, values);
+
+        if (result) {
+            console.log('성공');
+            //res.write('<h2>회원가입 성공</h2>');
+            res.cookie('id', id);
+            return res.redirect('/');
+        } else {
+            console.log('실패');
+            //res.write('<h2>회원가입 실패</h2>');
+            return res.redirect('/');
+        }
+    }
+});
+app.get('/board/signup', async function (req, res) {
+    return res.render('board_signup.html');
 });
 
 app.listen(3000, function () {
